requirements.txt
torch
torchvision
streamlit
scikit-learn
pillow